refactor: OrderStatus를 gradle api dependency를 사용해 common 모듈에서 domain 모듈로 이동
