+++ build.gradle
@@ -14,6 +14,7 @@ subprojects {
     apply plugin: 'org.springframework.boot'
     apply plugin: 'io.spring.dependency-management'
     apply plugin: 'java'
+    apply plugin: 'java-library'
 
     group = 'camp.nextstep.edu'
     version = '0.0.1-SNAPSHOT'
