+++ module-api/build.gradle
@@ -8,7 +8,7 @@ jar {
 }
 
 dependencies {
-    implementation project(":module-domain")
+    api project(":module-domain")
     implementation project(":module-common")
 
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
