+++ module-api/src/main/java/kitchenpos/order/dto/OrderRequest.java
@@ -1,6 +1,7 @@
 package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderStatus;
+
+import kitchenpos.order.domain.OrderStatus;
 
 import java.util.List;
