+++ module-api/src/main/java/kitchenpos/order/dto/OrderResponse.java
@@ -1,8 +1,8 @@
 package kitchenpos.order.dto;
 
-import kitchenpos.domain.OrderStatus;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
+import kitchenpos.order.domain.OrderStatus;
 
 import java.time.LocalDateTime;
 import java.util.List;
