+++ module-api/src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -1,9 +1,8 @@
 package kitchenpos.order.application;
 
-import kitchenpos.domain.OrderStatus;
-import kitchenpos.order.application.OrderService;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.dto.OrderResponse;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.order.dto.OrderLineItemRequest;
