+++ module-web/build.gradle
@@ -7,8 +7,7 @@ jar {
 }
 
 dependencies {
-    implementation project(':module-api')
-    implementation project(':module-common')
+    api project(':module-api')
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.flywaydb:flyway-core'
     runtimeOnly 'com.h2database:h2'
